requirements.txt
absl-py==0.14.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.5
cycler==0.10.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
h5py==3.4.0
idna==3.2
importlib-metadata==4.8.1
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.20.3
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.1
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-ignite==0.1.2
pytorch-ssim==0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.2.1
seaborn==0.11.2
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tqdm==4.62.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
yacs==0.1.8
zipp==3.5.0
